+++ src/main/java/com/storeware/calculator/infrastructure/output/ConsoleMessageOutputHandler.java
@@ -3,10 +3,12 @@ package com.storeware.calculator.infrastructure.output;
 import com.storeware.calculator.application.port.out.OutputStrategy;
 import com.storeware.calculator.util.ConsolePrinter;
 
+import java.math.BigDecimal;
+
 public class ConsoleMessageOutputHandler implements OutputStrategy {
 
     @Override
-    public void handleResult(double result) {
+    public void handleResult(BigDecimal result) {
         ConsolePrinter.printInfo(String.valueOf(result));
     }
 }
