+++ src/main/java/com/storeware/calculator/infrastructure/output/FileMessageOutputHandler.java
@@ -2,10 +2,12 @@ package com.storeware.calculator.infrastructure.output;
 
 import com.storeware.calculator.application.port.out.OutputStrategy;
 
+import java.math.BigDecimal;
+
 public class FileMessageOutputHandler implements OutputStrategy {
 
     @Override
-    public void handleResult(double result) {
-
+    public void handleResult(BigDecimal result) {
+        // TODO to implement
     }
 }
